backend/requirements.txt
aniso8601==9.0.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-WTF==1.2.2
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
passlib==1.7.4
pytz==2024.2
six==1.17.0
Werkzeug==3.1.3
WTForms==3.2.1
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Migrate==4.0.5
flask_socketio==5.5.0
flask_cors==5.0.0
flask_jwt_extended==4.7.1
flask_limiter==3.9.2
bcrypt==3.2.2